switch-bsp: only enable ddr200 by default on samsung branded cards
too many users of non-samsung branded cards have issues when ddr200 is enabled. this change to the uboot script makes is so ddr200 is disabled by default except on samsung branded cards even if the card says it supports it. if the user wants to enable ddr200 manually on such cards put `ddr200_enable=1` in the .ini